requirements.txt
pdfplumber>=0.10.0
ebooklib>=0.18
spacy>=3.7.0
beautifulsoup4>=4.12.0
chardet>=5.0.0
python-docx>=1.1.0
textblob>=0.17.1
